feat: reorganize things into files and made a simba to simba compiler*
This was growing too large and became hard to work with so the project was split into multiple files.
Because the end product has to be a single file, I've made "s2scompiler.simba" which is basically a simple "Simba to Simba compiler" from multiple files into a single one.